fix(weaver): move wheel handler to TimelinePanel to cover entire area
The wheel handler was only on the Ruler element, so scrolling on
stem lanes or empty areas triggered the macOS overscroll beep.
Moved zoom/pan wheel handling to the TimelinePanel wrapper so it
covers ruler + stem lanes + drop zone uniformly.
